+++ packages/json/src/jsonBlockPrompt.test.ts
@@ -11,7 +11,8 @@ describe("jsonBlockPrompt", () => {
         examples: [{ a: "example" }],
       }),
     ).toMatchInlineSnapshot(`
-      "You can respond with a simple component by wrapping JSON in 【】. The schema is:
+      "You can respond with a simple component by wrapping JSON in 【】.
+      The JSON schema is:
       {"type":"object","properties":{"a":{"type":"string"}},"required":["a"]}
 
       Examples: 
@@ -29,11 +30,31 @@ describe("jsonBlockPrompt", () => {
         examples: [{ a: [{ a: "a", b: "b" }] }],
       }),
     ).toMatchInlineSnapshot(`
-      "You can respond with a complex component by wrapping JSON in 【】. The schema is:
+      "You can respond with a complex component by wrapping JSON in 【】.
+      The JSON schema is:
       {"type":"object","properties":{"a":{"type":"array","items":{"type":"object","properties":{"a":{"type":"string"},"b":{"type":"string"}},"required":["a","b"],"additionalProperties":false}}},"required":["a"]}
 
       Examples: 
       【{"a":[{"a":"a","b":"b"}]}】"
+    `);
+  });
+
+  it("custom start and end chars", () => {
+    expect(
+      jsonBlockPrompt({
+        name: "complex",
+        schema: z.object({
+          a: z.array(z.object({ a: z.string(), b: z.string() })),
+        }),
+        examples: [{ a: [{ a: "a", b: "b" }] }],
+        options: { startChar: "z", endChar: "x" },
+      }),
+    ).toMatchInlineSnapshot(`
+      "You can respond with a complex component by wrapping JSON in zx. The schema is:
+      {"type":"object","properties":{"a":{"type":"array","items":{"type":"object","properties":{"a":{"type":"string"},"b":{"type":"string"}},"required":["a","b"],"additionalProperties":false}}},"required":["a"]}
+
+      Examples: 
+      z{"a":[{"a":"a","b":"b"}]}x"
     `);
   });
 });