+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llm-ui/markdown
 
+## 0.10.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @llm-ui/react@0.10.0
+
 ## 0.9.0
 
 ### Patch Changes